moved WindowedSpatialAverage to Diagnostics. It doesn't compile because it depends on OutputWriters, but I'll get there...
