+++ sandbox/windowed_spatial_avg.jl
@@ -37,46 +37,12 @@ set!(model, u=u0, v=v0, w=w0)
 
 
 #++++
-struct WindowedSpatialAverage{F, S, D}
-          field :: F
-   field_slicer :: S
-           dims :: D
-end
-WindowedSpatialAverage(field; dims, field_slicer=FieldSlicer()) = WindowedSpatialAverage(field, field_slicer, dims)
-
-using Oceananigans.OutputWriters: slice_parent
-using Statistics: mean
-function (wsa::WindowedSpatialAverage)(model)
-    compute!(wsa.field)
-    window = slice_parent(wsa.field_slicer, wsa.field)
-    return dropdims(mean(window, dims=wsa.dims), dims=wsa.dims)
-end
-
-using NCDatasets: defVar
-using Oceananigans.Fields: reduced_location
-import Oceananigans.OutputWriters: xdim, ydim, zdim, define_output_variable!
-
-function define_output_variable!(dataset, 
-                                 wtsa::Union{WindowedSpatialAverage, WindowedTimeAverage{<:WindowedSpatialAverage}}, 
-                                 name, array_type, compression, attributes, dimensions)
-    if wtsa isa WindowedSpatialAverage
-        wsa = wtsa
-    elseif wtsa isa WindowedTimeAverage
-        wsa = wtsa.operand
-    else
-        throw("Wrong type for windowed averaged")
-    end
-    LX, LY, LZ = reduced_location(location(wsa.field), dims=wsa.dims)
-    output_dims = tuple(xdim(LX)..., ydim(LY)..., zdim(LZ)...)
-    defVar(dataset, name, eltype(array_type), (output_dims..., "time"),
-           compression=compression, attrib=attributes)
-    return nothing
-end
 #----
 
 
 #++++
 using Oceananigans.Grids
+using Oceananigans.Diagnostics: WindowedSpatialAverage
 u, v, w = model.velocities
 slicer = FieldSlicer(j=Ny÷2+1:Ny)
 
@@ -91,11 +57,11 @@ progress(sim) = @printf("Iteration: %d, time: %s, Δt: %s\n",
                         prettytime(sim.model.clock.time),
                         prettytime(sim.Δt))
 
-simulation = Simulation(model, Δt=1second, iteration_interval=5, progress=progress, stop_iteration=10,)
+simulation = Simulation(model, Δt=1second, iteration_interval=5, progress=progress, stop_iteration=20,)
 
-wout = (;  Uw, U2w)
+wout = (; Uw, U2w)
 simulation.output_writers[:simple_output] = NetCDFOutputWriter(model, wout, 
-                                                               schedule = TimeInterval(10seconds),
+                                                               schedule = AveragedTimeInterval(10seconds),
                                                                filepath = "windowed_avg.nc", mode = "c")
 
 run!(simulation)
