+++ src/Diagnostics/Diagnostics.jl
@@ -5,7 +5,7 @@ export
     FieldMaximum,
     CFL, AdvectiveCFL, DiffusiveCFL,
     run_diagnostic!,
-    TimeInterval, IterationInterval, WallTimeInterval
+    TimeInterval, IterationInterval, WallTimeInterval, WindowedSpatialAverage
 
 using CUDA
 using Oceananigans
@@ -17,5 +17,6 @@ using Oceananigans: AbstractDiagnostic
 include("nan_checker.jl")
 include("field_maximum.jl")
 include("cfl.jl")
+include("windowed_spatial_average.jl")
 
 end
